refactor(envvars): split environment variable primitives from system extension
Move get-environment-variable and get-environment-variables from the system
extension to a new internal/extensions/envvars package. Environment variable
access is a configuration concern, not a system interface — this lets the
Console profile include envvars without pulling in exit/emergency-exit/etc.

The new primitives gain a virtual env map hook: if Namespace.EnvMap() is
non-nil, lookups read from the map directly (capability granted by providing
it). Otherwise they fall through to os.Getenv/os.Environ gated by the
authorizer. Namespace.envMap + EnvMap() are stubs in this change; the
WithEnvMap option wiring lands in Task 5.

Profile integration:
- envvars.Extension added to Console, ConsoleWithLoad, Small, KitchenSink
- AllExtensions(), bootstrap allExtensions also include envvars
- extension_consistency_test expected list updated
- TestAuthorizer_DenyBlocksEnvVar now imports envvars instead of system
