+++ extensions/system/prim_system_test.go
@@ -113,79 +113,6 @@ func TestCommandLine(t *testing.T) {
 	})
 }
 
-func TestGetEnvironmentVariable(t *testing.T) {
-	c := qt.New(t)
-	engine := newEngine(t)
-
-	t.Run("existing variable", func(t *testing.T) {
-		t.Setenv("WILE_TEST_VAR", "hello")
-		result := eval(t, engine, `(get-environment-variable "WILE_TEST_VAR")`)
-		c.Assert(result.Internal(), valuestest.SchemeEquals, values.NewString("hello"))
-	})
-
-	t.Run("nonexistent variable returns false", func(t *testing.T) {
-		result := eval(t, engine, `(get-environment-variable "WILE_NONEXISTENT_VAR_12345")`)
-		c.Assert(result.Internal(), qt.Equals, values.FalseValue)
-	})
-
-	t.Run("empty string value", func(t *testing.T) {
-		t.Setenv("WILE_TEST_EMPTY", "")
-		result := eval(t, engine, `(get-environment-variable "WILE_TEST_EMPTY")`)
-		c.Assert(result.Internal(), valuestest.SchemeEquals, values.NewString(""))
-	})
-
-	t.Run("value with equals sign", func(t *testing.T) {
-		t.Setenv("WILE_TEST_EQUALS", "a=b=c")
-		result := eval(t, engine, `(get-environment-variable "WILE_TEST_EQUALS")`)
-		c.Assert(result.Internal(), valuestest.SchemeEquals, values.NewString("a=b=c"))
-	})
-
-	t.Run("wrong argument type", func(t *testing.T) {
-		evalExpectError(t, engine, `(get-environment-variable 42)`)
-	})
-
-	t.Run("wrong argument count", func(t *testing.T) {
-		evalExpectError(t, engine, `(get-environment-variable)`)
-	})
-}
-
-func TestGetEnvironmentVariables(t *testing.T) {
-	c := qt.New(t)
-	engine := newEngine(t)
-
-	t.Run("returns a list", func(t *testing.T) {
-		result := eval(t, engine, `(list? (get-environment-variables))`)
-		c.Assert(result.Internal(), qt.Equals, values.TrueValue)
-	})
-
-	t.Run("contains known variable", func(t *testing.T) {
-		t.Setenv("WILE_TEST_ENVLIST", "found")
-		result := eval(t, engine, `
-			(let loop ((vars (get-environment-variables)))
-			  (cond
-			    ((null? vars) #f)
-			    ((string=? (car (car vars)) "WILE_TEST_ENVLIST")
-			     (cdr (car vars)))
-			    (else (loop (cdr vars)))))
-		`)
-		c.Assert(result.Internal(), valuestest.SchemeEquals, values.NewString("found"))
-	})
-
-	t.Run("entries are pairs", func(t *testing.T) {
-		result := eval(t, engine, `
-			(let ((vars (get-environment-variables)))
-			  (if (null? vars)
-			      #t
-			      (pair? (car vars))))
-		`)
-		c.Assert(result.Internal(), qt.Equals, values.TrueValue)
-	})
-
-	t.Run("wrong argument count", func(t *testing.T) {
-		evalExpectError(t, engine, `(get-environment-variables 42)`)
-	})
-}
-
 func TestCurrentSecond(t *testing.T) {
 	c := qt.New(t)
 	engine := newEngine(t)
