+++ extensions/system/register.go
@@ -39,12 +39,6 @@ func addPrimitives(r *registry.Registry) error {
 			Doc: "Exits the process. With no argument or #t, exits with code 0. With #f, exits with code 1. With an integer, uses that code.", ParamNames: []string{"status"}, Category: "system"},
 		{Name: "emergency-exit", ParamCount: 1, IsVariadic: true, Impl: PrimEmergencyExit,
 			Doc: "Exits the process immediately without running cleanup. Same argument handling as exit.", ParamNames: []string{"status"}, Category: "system"},
-		{Name: "get-environment-variable", ParamCount: 1, Impl: PrimGetEnvironmentVariable,
-			Doc: "Returns the value of the named environment variable as a string, or #f if not set.", ParamNames: []string{"name"}, Category: "system",
-			Keywords: []string{"getenv", "env var", "environment lookup"}},
-		{Name: "get-environment-variables", Impl: PrimGetEnvironmentVariables,
-			Doc: "Returns all environment variables as an association list of (name . value) pairs.", Category: "system",
-			Keywords: []string{"environ", "env vars", "all environment"}},
 		{Name: "current-second", Impl: PrimCurrentSecond,
 			Doc: "Returns the current time as seconds since the Unix epoch as an inexact real number.", Category: "system",
 			Keywords: []string{"time now", "Unix timestamp", "epoch", "wall clock"}},
