+++ internal/bootstrap/environment_tiny.go
@@ -43,6 +43,7 @@ import (
 	"github.com/aalpar/wile/extensions/system"
 	"github.com/aalpar/wile/extensions/threads"
 	"github.com/aalpar/wile/internal/extensions/all"
+	"github.com/aalpar/wile/internal/extensions/envvars"
 	ioext "github.com/aalpar/wile/internal/extensions/io"
 	nsext "github.com/aalpar/wile/internal/extensions/namespace"
 	"github.com/aalpar/wile/internal/parser"
@@ -76,6 +77,7 @@ var allExtensions = []registry.Extension{
 	all.Extension,
 	system.Extension,
 	process.Extension,
+	envvars.Extension,
 }
 
 // initializeEnvironmentWithRegistry is the shared initialization sequence for environment creation.
